+++ bxbot-exchanges/src/main/java/com/gazbert/bxbot/exchanges/GeminiExchangeAdapter.java
@@ -29,6 +29,7 @@ import com.gazbert.bxbot.exchange.api.ExchangeConfig;
 import com.gazbert.bxbot.exchange.api.OptionalConfig;
 import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderBookImpl;
 import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderImpl;
+import com.gazbert.bxbot.exchanges.trading.api.impl.OpenOrderImpl;
 import com.gazbert.bxbot.trading.api.*;
 import com.google.common.base.MoreObjects;
 import com.google.gson.Gson;
@@ -353,7 +354,7 @@ public final class GeminiExchangeAdapter extends AbstractExchangeAdapter impleme
                                 "Unrecognised order type received in getYourOpenOrders(). Value: " + geminiOpenOrder.type);
                 }
 
-                final OpenOrder order = new OpenOrder(
+                final OpenOrder order = new OpenOrderImpl(
                         Long.toString(geminiOpenOrder.order_id),
                         Date.from(Instant.ofEpochMilli(geminiOpenOrder.timestampms)),
                         marketId,
