+++ bxbot-exchanges/src/main/java/com/gazbert/bxbot/exchanges/HuobiExchangeAdapter.java
@@ -29,6 +29,7 @@ import com.gazbert.bxbot.exchange.api.ExchangeConfig;
 import com.gazbert.bxbot.exchange.api.OptionalConfig;
 import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderBookImpl;
 import com.gazbert.bxbot.exchanges.trading.api.impl.MarketOrderImpl;
+import com.gazbert.bxbot.exchanges.trading.api.impl.OpenOrderImpl;
 import com.gazbert.bxbot.trading.api.*;
 import com.google.common.base.MoreObjects;
 import com.google.gson.*;
@@ -377,7 +378,7 @@ public final class HuobiExchangeAdapter extends AbstractExchangeAdapter implemen
                                     "Unrecognised order type received in getYourOpenOrders(). Value: " + openOrder.type);
                     }
 
-                    final OpenOrder order = new OpenOrder(
+                    final OpenOrder order = new OpenOrderImpl(
                             Long.toString(openOrder.id),
                             new Date(openOrder.order_time),
                             marketId,
